refactor(gateway): merge Message, P2PMessage and P2PMessageV2 into one concrete message type
将 gateway 消息的三层继承（Message 虚接口 ← P2PMessage ← P2PMessageV2）合并为 libnetwork 中
具体、非多态的 Message 类：

- 版本差异数据化：原 P2PMessageV2 的 ttl 与扩展头（ttl/srcP2PNodeID/dstP2PNodeID）编解码
  并入 Message，按 version 字段内部分支，wire 格式逐字节不变；同时修正了旧 P2PMessage
  在 version > V0 时编码出"声称 V2 却缺少扩展头"的不一致帧的问题
- 移除全部虚函数：getter 可内联，消除虚表间接跳转；MessageFactory 同步具体化
- 依赖方向清理：压缩常数（c_compressThreshold/c_zstdCompressLevel）下沉至
  libnetwork/Common.h，libnetwork 不再反向 include libp2p 头文件
- dynamic_pointer_cast 全部降为 static_pointer_cast，删除与类型别名冲突的前向声明
- 测试缝由运行时多态改为真实帧：构造合法 V0 帧与畸形帧（非法版本、扩展头截断）覆盖
  读循环的 decode 成功、错误与异常分支，替代原 FakeP2PMessage/DecodeErrorMessage 等替身类

验证：全项目编译通过，gateway 146 个单元测试全部通过。
