+++ bcos-gateway/bcos-gateway/GatewayFactory.cpp
@@ -17,10 +17,10 @@
 #include "bcos-gateway/libnetwork/Host.h"
 #include "bcos-gateway/libnetwork/PeerBlackWhitelistInterface.h"
 #include "bcos-gateway/libnetwork/PeerBlacklist.h"
+#include "bcos-gateway/libnetwork/Message.h"
 #include "bcos-gateway/libnetwork/PeerWhitelist.h"
 #include "bcos-gateway/libnetwork/Session.h"
 #include "bcos-gateway/libnetwork/SessionCallback.h"
-#include "bcos-gateway/libp2p/P2PMessageV2.h"
 #include "bcos-gateway/libp2p/Service.h"
 #include "bcos-gateway/libp2p/ServiceV2.h"
 #include "bcos-gateway/libp2p/router/RouterTableImpl.h"
@@ -692,7 +692,7 @@ std::shared_ptr<Service> GatewayFactory::buildService(const GatewayConfig::Ptr& 
     asioInterface->setType(ASIOInterface::ASIO_TYPE::SSL);
 
     // Message Factory
-    auto messageFactory = std::make_shared<P2PMessageFactoryV2>();
+    auto messageFactory = std::make_shared<MessageFactory>();
     auto nodeIDHash = _config->calculateShortNodeID(pubHex);
     P2PInfo selfInfo(nodeIDHash, pubHex);
     // Session Factory
@@ -815,7 +815,7 @@ std::shared_ptr<Gateway> GatewayFactory::buildGateway(GatewayConfig::Ptr _config
                 service->registerHandlerByMsgType(GatewayMessageType::AMOPMessageType,
                     [](const bcos::gateway::NetworkException& _e,
                         const bcos::gateway::P2PSession::Ptr& session,
-                        const std::shared_ptr<bcos::gateway::P2PMessage>& message) {
+                        const std::shared_ptr<bcos::gateway::Message>& message) {
                         // 只读模式下, 不处理其它节点的amop消息
                         // In read-only mode, AMOP messages from other nodes are not processed
                         return;
