+++ bcos-gateway/bcos-gateway/gateway/LocalRouterTable.h
@@ -23,13 +23,13 @@
 #include "bcos-crypto/interfaces/crypto/KeyInterface.h"
 #include "bcos-framework/gateway/GatewayInterface.h"
 #include "bcos-framework/multigroup/GroupInfo.h"
+#include "bcos-gateway/libnetwork/Message.h"
 #include <bcos-task/Task.h>
 #include <memory>
 namespace bcos
 {
 namespace gateway
 {
-class P2PMessage;
 
 class LocalRouterTable
 {
@@ -59,7 +59,7 @@ public:
     GroupNodeListType nodeList() const;
 
     bool broadcastMsg(uint16_t _nodeType, const std::string& _groupID, uint16_t _moduleID,
-        bcos::crypto::NodeIDPtr _srcNodeID, std::shared_ptr<P2PMessage> _msg) const;
+        bcos::crypto::NodeIDPtr _srcNodeID, std::shared_ptr<Message> _msg) const;
 
     task::Task<bcos::Error::Ptr> sendMessage(std::string _groupID,
         bcos::crypto::NodeIDPtr _srcNodeID, bcos::crypto::NodeIDPtr _dstNodeID,
