+++ bcos-gateway/bcos-gateway/libp2p/P2PSession.cpp
@@ -5,9 +5,8 @@
 
 #include "bcos-gateway/libp2p/P2PSession.h"
 #include "bcos-gateway/libnetwork/ASIOInterface.h"
+#include "bcos-gateway/libnetwork/Message.h"
 #include "bcos-gateway/libp2p/Common.h"
-#include "bcos-gateway/libp2p/P2PMessage.h"
-#include "bcos-gateway/libp2p/P2PMessageV2.h"
 #include "bcos-gateway/libp2p/Service.h"
 #include "bcos-utilities/Common.h"
 #include <bcos-task/Wait.h>
@@ -135,7 +134,7 @@ void P2PSession::heartBeat()
             try
             {
                 task::wait([](std::shared_ptr<P2PSession> _self) -> task::Task<void> {
-                    P2PMessageV2 message;
+                    Message message;
                     message.setPacketType(GatewayMessageType::Heartbeat);
                     ::ranges::any_view<bytesConstRef> emptyPayloads;
                     co_await _self->fastSendP2PMessage(
@@ -169,7 +168,7 @@ void P2PSession::heartBeat()
 }
 
 bcos::task::Task<Message::Ptr> P2PSession::fastSendP2PMessage(
-    P2PMessage& message, ::ranges::any_view<bytesConstRef> payloads, Options options)
+    Message& message, ::ranges::any_view<bytesConstRef> payloads, Options options)
 {
     if (!m_session || !m_session->active()) [[unlikely]]
     {
@@ -187,7 +186,7 @@ bcos::task::Task<Message::Ptr> P2PSession::fastSendP2PMessage(
     // Note: m_protocolInfo be setted when create P2PSession
     service->resetP2pID(message, (ProtocolVersion)m_protocolInfo->version());
     // the p2p message version must match the negotiated protocol version of this session: the
-    // encodeHeaderImpl of P2PMessageV2 only encodes the ttl/src/dst routing fields for version > V0,
+    // encodeHeaderImpl of Message only encodes the ttl/src/dst routing fields for version > V0,
     // so sending with the default (V0) version would silently drop the V2 routing fields and break
     // multi-hop forwarding through ServiceV2 router tables
     message.setVersion((uint16_t)m_protocolInfo->version());
