+++ bcos-gateway/test/unittests/FIB183_DecodedFieldBoundsTest.cpp
@@ -19,8 +19,7 @@
  */
 
 #include "bcos-gateway/Gateway.h"
-#include "bcos-gateway/libp2p/P2PMessage.h"
-#include "bcos-gateway/libp2p/P2PMessageV2.h"
+#include "bcos-gateway/libnetwork/Message.h"
 #include "bcos-gateway/libp2p/ServiceV2.h"
 #include "bcos-gateway/libp2p/router/RouterTableImpl.h"
 #include "bcos-utilities/testutils/TestPromptFixture.h"
@@ -44,7 +43,7 @@ public:
     void stop() override {}
 
     void callOnReceiveP2PMessage(
-        NetworkException const& _e, P2PSession::Ptr _session, std::shared_ptr<P2PMessage> _msg)
+        NetworkException const& _e, P2PSession::Ptr _session, std::shared_ptr<Message> _msg)
     {
         onReceiveP2PMessage(_e, _session, _msg);
     }
@@ -60,21 +59,21 @@ public:
       : ServiceV2(_info, std::move(_factory), _ioContext)
     {}
     void callOnReceiveRouterSeq(
-        NetworkException _error, std::shared_ptr<P2PSession> _session, P2PMessage::Ptr _message)
+        NetworkException _error, std::shared_ptr<P2PSession> _session, Message::Ptr _message)
     {
         onReceiveRouterSeq(std::move(_error), std::move(_session), std::move(_message));
     }
 };
 
-// Fix A: a P2PMessage whose decoded options carry zero dstNodeIDs must not be
+// Fix A: a Message whose decoded options carry zero dstNodeIDs must not be
 // indexed (dstNodeIDs[0] on an empty vector is OOB). The guard drops it before
 // touching m_gatewayNodeManager (which is null here), so no crash occurs.
 BOOST_AUTO_TEST_CASE(EmptyDstNodeIDsIsDropped)
 {
-    // Build a real P2PMessage with a non-zero moduleID (so the front-message
+    // Build a real Message with a non-zero moduleID (so the front-message
     // moduleID-decode branch is skipped) and an empty dstNodeIDs list.
-    auto factory = std::make_shared<P2PMessageFactory>();
-    auto msg = std::static_pointer_cast<P2PMessage>(factory->buildMessage());
+    auto factory = std::make_shared<MessageFactory>();
+    auto msg = std::static_pointer_cast<Message>(factory->buildMessage());
     msg->setPacketType(GatewayMessageType::PeerToPeerMessage);
 
     P2PMessageOptions options;
@@ -122,11 +121,11 @@ BOOST_AUTO_TEST_CASE(ShortRouterSeqPayloadIsDropped)
     auto routerTableFactory = std::make_shared<RouterTableFactoryImpl>();
     boost::asio::io_context ioContext;
     auto service = std::make_shared<FakeServiceV2FIB183>(selfInfo, routerTableFactory, ioContext);
-    auto factory = std::make_shared<P2PMessageFactoryV2>();
+    auto factory = std::make_shared<MessageFactory>();
 
     for (size_t len = 0; len < sizeof(uint32_t); ++len)
     {
-        auto msg = std::static_pointer_cast<P2PMessage>(factory->buildMessage());
+        auto msg = std::static_pointer_cast<Message>(factory->buildMessage());
         msg->setPacketType(GatewayMessageType::RouterTableSyncSeq);
         if (len > 0)
         {
