+++ bcos-gateway/test/unittests/FIB186_BroadcastLockOrderTest.cpp
@@ -34,7 +34,7 @@
  */
 
 #include "bcos-framework/gateway/GatewayTypeDef.h"
-#include "bcos-gateway/libp2p/P2PMessage.h"
+#include "bcos-gateway/libnetwork/Message.h"
 #include "bcos-gateway/libp2p/P2PSession.h"
 #include "bcos-gateway/libp2p/Service.h"
 #include "bcos-task/Wait.h"
@@ -70,7 +70,7 @@ public:
     // virtual is called for each session; a probe thread tries to take x_sessions EXCLUSIVELY: it
     // fails iff the broadcasting thread still holds it. No throw needed -- broadcastMessageToAll
     // touches no host; we just probe the lock state and stop.
-    task::Task<Message::Ptr> sendMessageByNodeID(P2pID /*nodeID*/, P2PMessage& /*header*/,
+    task::Task<Message::Ptr> sendMessageByNodeID(P2pID /*nodeID*/, Message& /*header*/,
         ::ranges::any_view<bytesConstRef> /*payloads*/, Options /*options*/) override
     {
         bool acquiredExclusive = false;
@@ -108,8 +108,8 @@ BOOST_AUTO_TEST_CASE(BroadcastDoesNotHoldSessionsLockWhileSending)
     auto service = std::make_shared<BroadcastProbeService>(selfInfo);
     service->arm();
 
-    auto message = std::make_shared<P2PMessage>();
-    task::wait([](std::shared_ptr<BroadcastProbeService> _service, P2PMessage::Ptr _message)
+    auto message = std::make_shared<Message>();
+    task::wait([](std::shared_ptr<BroadcastProbeService> _service, Message::Ptr _message)
                    -> task::Task<void> {
         co_await _service->broadcastMessageToAll(
             _message, ::ranges::views::single(_message->payload()), Options{});
