+++ bcos-gateway/test/unittests/FIB186_BulkDisconnectReactorTest.cpp
@@ -44,8 +44,8 @@
 #include "bcos-crypto/hash/Keccak256.h"
 #include "bcos-gateway/libnetwork/ASIOInterface.h"
 #include "bcos-gateway/libnetwork/Host.h"
+#include "bcos-gateway/libnetwork/Message.h"
 #include "bcos-gateway/libnetwork/Session.h"
-#include "bcos-gateway/libp2p/P2PMessage.h"
 #include "bcos-utilities/IOServicePool.h"
 #include "bcos-utilities/testutils/TestPromptFixture.h"
 #include <chrono>
@@ -140,7 +140,7 @@ BOOST_AUTO_TEST_CASE(TeardownFloodMustNotStarveMessageDelivery)
     // process-global TBB control, so "the flood would swamp the delivery reactor if it landed
     // there" is deterministic and independent of the host core count.
     auto hashImpl = std::make_shared<Keccak256>();
-    auto messageFactory = std::make_shared<P2PMessageFactory>();
+    auto messageFactory = std::make_shared<MessageFactory>();
     auto fakeAsio = std::make_shared<FakeASIO_Reactor>();
     auto fakeHost = std::make_shared<FakeHost_Reactor>(hashImpl, fakeAsio, messageFactory);
