+++ bcos-gateway/test/unittests/amop/AMOPSendMessagePathTest.cpp
@@ -23,8 +23,8 @@
 #include "bcos-gateway/libamop/AMOPImpl.h"
 #include "bcos-gateway/libamop/AMOPMessage.h"
 #include "bcos-gateway/libamop/TopicManager.h"
+#include "bcos-gateway/libnetwork/Message.h"
 #include "bcos-gateway/libp2p/P2PInterface.h"
-#include "bcos-gateway/libp2p/P2PMessageV2.h"
 #include "bcos-utilities/IOServicePool.h"
 #include "bcos-utilities/testutils/TestPromptFixture.h"
 
@@ -60,7 +60,7 @@ bcos::bytes encodeAMOPResponse(uint16_t _status, std::string const& _data)
 Message::Ptr buildP2PResponse(
     bcos::bytes _payload, uint16_t _packetType = GatewayMessageType::AMOPMessageType)
 {
-    auto message = std::make_shared<P2PMessageV2>();
+    auto message = std::make_shared<Message>();
     message->setPacketType(_packetType);
     message->setPayload(std::move(_payload));
     return message;
@@ -86,7 +86,7 @@ struct AMOPSendFixture
         // TopicManager::queryNodeIDsByTopic only returns reachable nodes
         When(Method(networkMock, isReachable)).AlwaysReturn(true);
         When(Method(networkMock, messageFactory)).AlwaysDo([]() -> std::shared_ptr<MessageFactory> {
-            return std::make_shared<P2PMessageFactoryV2>();
+            return std::make_shared<MessageFactory>();
         });
 
         network = P2PInterface::Ptr(&networkMock.get(), [](P2PInterface*) {});
@@ -154,7 +154,7 @@ BOOST_AUTO_TEST_CASE(test_allCandidatesFail)
 
     auto attempts = fixture.attempts;
     When(Method(fixture.networkMock, sendMessageByNodeID))
-        .AlwaysDo([attempts](P2pID nodeID, P2PMessage&, ::ranges::any_view<bytesConstRef>,
+        .AlwaysDo([attempts](P2pID nodeID, Message&, ::ranges::any_view<bytesConstRef>,
                       Options) -> task::Task<Message::Ptr> {
             attempts->push_back(nodeID);
             if (!nodeID.empty())
@@ -197,7 +197,7 @@ BOOST_AUTO_TEST_CASE(test_retrySucceedsAfterNetworkException)
 
     auto attempts = fixture.attempts;
     When(Method(fixture.networkMock, sendMessageByNodeID))
-        .AlwaysDo([attempts, expectedPayload](P2pID nodeID, P2PMessage&,
+        .AlwaysDo([attempts, expectedPayload](P2pID nodeID, Message&,
                       ::ranges::any_view<bytesConstRef>, Options) -> task::Task<Message::Ptr> {
             attempts->push_back(nodeID);
             if (attempts->size() == 1)
@@ -236,7 +236,7 @@ BOOST_AUTO_TEST_CASE(test_nullResponseRetriesNextNode)
     auto expectedPayload = encodeAMOPResponse(0, "ok");
     auto attempts = fixture.attempts;
     When(Method(fixture.networkMock, sendMessageByNodeID))
-        .AlwaysDo([attempts, expectedPayload](P2pID nodeID, P2PMessage&,
+        .AlwaysDo([attempts, expectedPayload](P2pID nodeID, Message&,
                       ::ranges::any_view<bytesConstRef>, Options) -> task::Task<Message::Ptr> {
             attempts->push_back(nodeID);
             if (attempts->size() == 1)
@@ -269,7 +269,7 @@ BOOST_AUTO_TEST_CASE(test_malformedAMOPResponseFailsWithoutRetry)
 
     auto attempts = fixture.attempts;
     When(Method(fixture.networkMock, sendMessageByNodeID))
-        .AlwaysDo([attempts](P2pID nodeID, P2PMessage&, ::ranges::any_view<bytesConstRef>,
+        .AlwaysDo([attempts](P2pID nodeID, Message&, ::ranges::any_view<bytesConstRef>,
                       Options) -> task::Task<Message::Ptr> {
             attempts->push_back(nodeID);
             // AMOPMessage::decode needs at least the 6-byte header
@@ -300,7 +300,7 @@ BOOST_AUTO_TEST_CASE(test_sendUsesFiniteResponseTimeout)
     auto expectedPayload = encodeAMOPResponse(0, "ok");
     auto observedOptions = std::make_shared<std::vector<Options>>();
     When(Method(fixture.networkMock, sendMessageByNodeID))
-        .AlwaysDo([observedOptions, expectedPayload](P2pID, P2PMessage&,
+        .AlwaysDo([observedOptions, expectedPayload](P2pID, Message&,
                       ::ranges::any_view<bytesConstRef>, Options options)
                       -> task::Task<Message::Ptr> {
             observedOptions->push_back(options);
